CHANGED: splitter Main.main()
 * Moved gui.LaunchFrameHelpers to main.MainHelpers
